Crate salvo_core

The core lib of Savlo web server framework. Read more: https://salvo.rs

Re-exports

pub use hyper;

pub use self::catcher::Catcher;

pub use self::catcher::CatcherImpl;

pub use self::http::Request;

pub use self::http::Response;

pub use self::listener::Listener;

pub use self::writer::Piece;

pub use self::writer::Writer;

Modules

addr

addr module

catcher

Catcher tarit and it’s defalut implement CatcherImpl.

fs

Filesystem module

http

Http module

listener

Listener trait and it’s implements.

prelude

A list of things that automatically imports into application use salvo.

routing

Routing and filters Router can route http requests to different handlers.

writer

Writer trait and it’s implements.

Structs

Depot

Depot if for store temp data of current request. Each handler can read or write data to it.

Router

Router struct is used for route request to different handlers.

Server

HTTP Server

Service

Service http request.

Enums

Error

Errors that can happen inside salvo.

Traits

Handler

Handler is used for handle Request.

Functions

run

If you don’t want to include tokio in your project directly, you can use this function to run server.

run_with_threads

If you don’t want to include tokio in your project directly, you can use this function to run server.

Type Definitions

Result

Result type wich has salvo::Error as it’s error type.

Attribute Macros

async_trait

fn_handler

fn_handler is a pro macro to help create Handler from function easily.