Struct salvo_core::fs::NamedFile

pub struct NamedFile { /* private fields */ }

A file with an associated name.

Implementations

impl NamedFile

pub fn builder(path: impl Into<PathBuf>) -> NamedFileBuilder

Create new NamedFileBuilder.

pub async fn open(path: impl Into<PathBuf>) -> Result<NamedFile>

Attempts to open a file in read-only mode.

Examples

let file = NamedFile::open("foo.txt").await;

pub async fn send_file(
path: impl Into<PathBuf>,
req: &mut Request,
res: &mut Response
)

Attempts to send a file. If file not exists, not found error will occur.

pub fn file(&self) -> &File

Returns reference to the underlying File object.

pub fn path(&self) -> &Path

Retrieve the path of this file.

pub fn content_type(&self) -> &Mime

Get content type value.

pub fn set_content_type(&mut self, content_type: Mime)

Set the MIME Content-Type for serving this file. By default the Content-Type is inferred from the filename extension.

pub fn content_disposition(&self) -> &HeaderValue

Get Content-Disposition value.

pub fn set_content_disposition(&mut self, content_disposition: HeaderValue)

Set the Content-Disposition for serving this file. This allows changing the inline/attachment disposition as well as the filename sent to the peer.

By default the disposition is inline for text, image, and video content types, and attachment otherwise, and the filename is taken from the path provided in the open method after converting it to UTF-8 using to_string_lossy.

pub fn disable_content_disposition(&mut self)

Disable Content-Disposition header.

By default Content-Disposition` header is enabled.

pub fn content_encoding(&self) -> Option<&HeaderValue>

Get content encoding value reference.

pub fn set_content_encoding(&mut self, content_encoding: HeaderValue)

Set content encoding for serving this file

pub fn etag(&self) -> Option<ETag>

Get ETag value.

pub fn use_etag(&mut self, value: bool)

Specifies whether to use ETag or not.

Default is true.

pub fn last_modified(&self) -> Option<SystemTime>

GEt last_modified value.

pub fn use_last_modified(&mut self, value: bool)

Specifies whether to use Last-Modified or not.

Default is true.

pub async fn send(self, req: &mut Request, res: &mut Response)

Consume self and send content to Response.