Module salvo_core::http::form

Structs

FilePart
A file that is to be inserted into a multipart/* or alternatively an uploaded file that was received as part of multipart/* parsing.

FormData
The extracted text fields and uploaded files from a multipart/form-data request.

Functions

read_form_data
Parse MIME multipart/form-data information from a stream as a FormData.