Struct salvo_core::http::request::Request

pub struct Request { /* private fields */ }

Represents an HTTP request.

Stores all the properties of the client’s request.

Implementations

impl Request

pub fn new() -> Request

Creates a new blank Request

pub fn uri(&self) -> &Uri

Returns a reference to the associated URI.

Examples

let req = Request::default();
assert_eq!(*req.uri(), *"/");

pub fn uri_mut(&mut self) -> &mut Uri

Returns a mutable reference to the associated URI.

Examples

let mut req: Request= Request::default();
*req.uri_mut() = "/hello".parse().unwrap();
assert_eq!(*req.uri(), *"/hello");

pub fn method(&self) -> &Method

Returns a reference to the associated HTTP method.

Examples

let req = Request::default();
assert_eq!(*req.method(), Method::GET);

pub fn method_mut(&mut self) -> &mut Method

Returns a mutable reference to the associated HTTP method.

Examples

let mut request: Request = Request::default();
*request.method_mut() = Method::PUT;
assert_eq!(*request.method(), Method::PUT);

pub fn version(&self) -> Version

Returns the associated version.

pub fn version_mut(&mut self) -> &mut Version

Returns a mutable reference to the associated version.

pub fn remote_addr(&self) -> Option<&SocketAddr>

Get request remote address.

pub fn headers(&self) -> &HeaderMap

Returns a reference to the associated header field map.

Examples

let req = Request::default();
assert!(req.headers().is_empty());

pub fn headers_mut(&mut self) -> &mut HeaderMap<HeaderValue>

Returns a mutable reference to the associated header field map.

Examples

let mut req: Request = Request::default();
req.headers_mut().insert(HOST, HeaderValue::from_static("world"));
assert!(!req.headers().is_empty());

pub fn header<T>(&self, key: &str) -> Option<T> where
T: FromStr,

Get header with supplied name and try to parse to a ‘T’, returns None if failed or not found.

pub fn body(&self) -> Option<&Body>

Returns a reference to the associated HTTP body.

Examples

let req = Request::default();
assert!(req.body().is_some());

pub fn body_mut(&mut self) -> Option<&mut Body>

Returns a mutable reference to the associated HTTP body.

pub fn take_body(&mut self) -> Option<Body>

Take body form the request, and set the body to None in the request.

pub fn extensions(&self) -> &Extensions

Returns a reference to the associated extensions.

Examples

let req = Request::default();
assert!(req.extensions().get::<i32>().is_none());

pub fn extensions_mut(&mut self) -> &mut Extensions

Returns a mutable reference to the associated extensions.

Examples

let mut req = Request::default();
req.extensions_mut().insert("hello");
assert_eq!(req.extensions().get(), Some(&"hello"));

pub fn accept(&self) -> Vec<Mime>

Get accept.

pub fn first_accept(&self) -> Option<Mime>

Get first accept.

pub fn content_type(&self) -> Option<Mime>

Get content type.

pub fn cookies(&self) -> &CookieJar

Get CookieJar reference.

pub fn cookies_mut(&mut self) -> &mut CookieJar

Get CookieJar mutable reference.

pub fn cookie<T>(&self, name: T) -> Option<&Cookie<'static>> where
T: AsRef<str>,

Get Cookie from cookies.

pub fn params(&self) -> &HashMap<String, String>

Get params reference.

pub fn params_mut(&mut self) -> &mut HashMap<String, String>

Get params mutable reference.

pub fn param<T>(&self, key: &str) -> Option<T> where
T: FromStr,

Get param value from params.

pub fn queries(&self) -> &MultiMap<String, String>

Get queries reference.

pub fn query<F>(&self, key: &str) -> Option<F> where
F: FromStr,

Get query value from queries.

pub async fn form<F>(&mut self, key: &str) -> Option<F> where
F: FromStr,

Get field data from form.

pub async fn file(&mut self, key: &str) -> Option<&FilePart>

Get FilePart reference from request.

pub async fn first_file(&mut self) -> Option<&FilePart>

Get FilePart reference from request.

pub async fn files(&mut self, key: &str) -> Option<&Vec<FilePart>>

Get FilePart list reference from request.

pub async fn all_files(&mut self) -> Vec<&FilePart>

Get FilePart list reference from request.

pub async fn form_or_query<F>(&mut self, key: &str) -> Option<F> where
F: FromStr,

Get value from form first if not found then get from query.

pub async fn query_or_form<F>(&mut self, key: &str) -> Option<F> where
F: FromStr,

Get value from query first if not found then get from form.

pub async fn payload(&mut self) -> Result<&Vec<u8>, ParseError>

Get request payload.

pub async fn form_data(&mut self) -> Result<&FormData, ParseError>

Get FormData reference from request.

pub async fn parse_data<'de, T>(
&'de mut self,
sources: BitFlags<ParseSource>
) -> Result<T, ParseError> where
T: Deserialize<'de>,

Read url params as type T from request’s different sources.

Returns error if the same key is appeared in different sources. This function will not handle if payload is json format, use [parse_json] to get typed json payload.

pub fn parse_params<'de, T>(&'de mut self) -> Result<T, ParseError> where
T: Deserialize<'de>,

Read url params as type T from request.

pub fn parse_queries<'de, T>(&'de mut self) -> Result<T, ParseError> where
T: Deserialize<'de>,

Read queries as type T from request.

pub fn parse_headers<'de, T>(&'de mut self) -> Result<T, ParseError> where
T: Deserialize<'de>,

Read headers as type T from request.

pub async fn parse_json<'de, T>(&'de mut self) -> Result<T, ParseError> where
T: Deserialize<'de>,

Read body as type T from request.

pub async fn parse_form<'de, T>(&'de mut self) -> Result<T, ParseError> where
T: Deserialize<'de>,

Read body as type T from request.

pub async fn parse_body<T>(&mut self) -> Result<T, ParseError> where
T: DeserializeOwned,

Read body as type T from request.