Module salvo_core::prelude

A list of things that automatically imports into application use salvo.

Re-exports

pub use crate::http::Request;

pub use crate::http::Response;

pub use crate::listener::UnixListener;

pub use crate::listener::JoinedListener;

pub use crate::listener::Listener;

pub use crate::listener::TcpListener;

pub use crate::routing::FlowCtrl;

pub use crate::writer::Json;

pub use crate::writer::Piece;

pub use crate::writer::Text;

pub use crate::writer::Writer;

Structs

Depot

Depot if for store temp data of current request. Each handler can read or write data to it.

Router

Router struct is used for route request to different handlers.

Server

HTTP Server

Service

Service http request.

StatusCode

An HTTP status code (status-code in RFC 7230 et al.).

StatusError

StatusError contains http error information.

Traits

Handler

Handler is used for handle Request.

Attribute Macros

async_trait

fn_handler

fn_handler is a pro macro to help create Handler from function easily.